# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Gateway API
  description: Agent 网关中台后端，提供账户、项目、服务管理和模型网关接口。
  version: 0.1.0
paths:
...
  /api/admin/v1/projects/{project_id}/resources/directories:
    get:
      tags:
      - Project Resources
      summary: 查询项目固定资源目录
      description: 返回固定目录分类；目录不可新增、重命名或删除。
      operationId: list_resource_directories_api_admin_v1_projects__project_id__resources_directories_get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResourceDirectoryResponse'
                title: Response List Resource Directories Api Admin V1 Projects  Project
                  Id  Resources Directories Get
        '401':
          description: 需要有效登录令牌。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '403':
          description: 无权访问该项目。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '404':
          description: 项目或资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/v1/projects/{project_id}/resources:
    get:
      tags:
      - Project Resources
      summary: 列出项目资源
      description: 按项目访问权限列出有效资源；默认返回六个固定目录的全部资源。
      operationId: list_resources_api_admin_v1_projects__project_id__resources_get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: resource_type
        in: query
        required: false
        schema:
          anyOf:
          - enum:
            - memory
            - template
            type: string
          - type: 'null'
          description: 固定资源类型。若提供则必须同时提供 category。
          title: Resource Type
        description: 固定资源类型。若提供则必须同时提供 category。
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - enum:
            - sessions
            - profiles
            - longterm
            - system
            - user
            - assistant
            type: string
          - type: 'null'
          description: 固定目录分类。
          title: Category
        description: 固定目录分类。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResourceResponse'
                title: Response List Resources Api Admin V1 Projects  Project Id  Resources
                  Get
        '401':
          description: 需要有效登录令牌。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '403':
          description: 无权访问该项目。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '404':
          description: 项目或资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Project Resources
      summary: 创建项目资源
      description: 创建目录内同名唯一的资源；需有项目 owner/editor 权限。
      operationId: create_resource_api_admin_v1_projects__project_id__resources_post
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceResponse'
        '401':
          description: 需要有效登录令牌。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '403':
          description: 无权访问该项目。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '404':
          description: 项目或资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '409':
          description: 名称冲突或资源版本已过期。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '422':
          description: 请求字段或固定目录无效。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
  /api/admin/v1/projects/{project_id}/resources/{resource_id}:
    get:
      tags:
      - Project Resources
      summary: 读取项目资源
      description: 同时按项目 ID 和资源 ID 查找，防止跨项目读取。
      operationId: get_resource_api_admin_v1_projects__project_id__resources__resource_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: resource_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Resource Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceResponse'
        '401':
          description: 需要有效登录令牌。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '403':
          description: 无权访问该项目。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '404':
          description: 项目或资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Project Resources
      summary: 更新项目资源
      description: 版本匹配时更新并递增版本；陈旧版本返回 409 version_conflict。
      operationId: update_resource_api_admin_v1_projects__project_id__resources__resource_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: resource_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Resource Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceResponse'
        '401':
          description: 需要有效登录令牌。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '403':
          description: 无权访问该项目。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '404':
          description: 项目或资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '409':
          description: 名称冲突或资源版本已过期。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '422':
          description: 请求字段或固定目录无效。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
    delete:
      tags:
      - Project Resources
      summary: 软删除项目资源
      description: 版本匹配时设置软删除时间；删除项从默认目录列表中隐藏。
      operationId: delete_resource_api_admin_v1_projects__project_id__resources__resource_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: resource_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Resource Id
      - name: expected_version
        in: query
        required: true
        schema:
          type: integer
          minimum: 1
          description: 最近读取的资源版本。
          title: Expected Version
        description: 最近读取的资源版本。
      responses:
        '204':
          description: Successful Response
        '401':
          description: 需要有效登录令牌。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '403':
          description: 无权访问该项目。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '404':
          description: 项目或资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '409':
          description: 名称冲突或资源版本已过期。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
        '422':
          description: 请求字段或固定目录无效。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceErrorResponse'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ResourceCreateRequest:
      properties:
        resource_type:
          type: string
          enum:
          - memory
          - template
          title: Resource Type
          description: 固定资源类型。
          examples:
          - template
        category:
          type: string
          enum:
          - sessions
          - profiles
          - longterm
          - system
          - user
          - assistant
          title: Category
          description: 固定目录分类。
          examples:
          - system
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
          description: 目录内唯一文件名，不含路径。
          examples:
          - support-agent
        content:
          type: string
          maxLength: 100000
          title: Content
          description: 用户主动保存的资源正文，最多 100,000 字符。
          examples:
          - 你是一个客服助手。请用{{language}}回答以下问题：{{question}}
      type: object
      required:
      - resource_type
      - category
      - name
      - content
      title: ResourceCreateRequest
      description: 新建固定目录内的项目资源。
      examples:
      - category: system
        content: 你是一个客服助手。请用{{language}}回答以下问题：{{question}}
        name: support-agent
        resource_type: template
    ResourceDirectoryResponse:
      properties:
        resource_type:
          type: string
          title: Resource Type
          description: 资源类型：memory 或 template。
        category:
          type: string
          title: Category
          description: 固定目录代码，不是文件系统路径。
        label:
          type: string
          title: Label
          description: 用于界面展示的目录名称。
      type: object
      required:
      - resource_type
      - category
      - label
      title: ResourceDirectoryResponse
      description: 固定目录中一项不可变分类。
    ResourceErrorDetail:
      properties:
        code:
          type: string
          title: Code
          description: 机器可读错误码，例如 version_conflict。
        message:
          type: string
          title: Message
          description: 可展示给用户的错误原因。
      type: object
      required:
      - code
      - message
      title: ResourceErrorDetail
      description: 资源 API 稳定错误码和用户可读原因。
    ResourceErrorResponse:
      properties:
        detail:
          $ref: '#/components/schemas/ResourceErrorDetail'
          description: 结构化错误详情。
      type: object
      required:
      - detail
      title: ResourceErrorResponse
      description: 资源 API 错误响应结构。
    ResourceResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
          description: 稳定资源 ID。
        project_id:
          type: string
          format: uuid
          title: Project Id
          description: 所属项目 ID。
        resource_type:
          type: string
          enum:
          - memory
          - template
          title: Resource Type
          description: 固定资源类型。
        category:
          type: string
          enum:
          - sessions
          - profiles
          - longterm
          - system
          - user
          - assistant
          title: Category
          description: 固定目录分类。
        name:
          type: string
          title: Name
          description: 文件名。
        content:
          type: string
          title: Content
          description: 资源正文；与网关调用日志分离。
        version:
          type: integer
          title: Version
          description: 每次成功修改或删除后递增的版本号。
        created_by:
          type: string
          format: uuid
          title: Created By
          description: 创建者用户 ID。
        updated_by:
          type: string
          format: uuid
          title: Updated By
          description: 最近修改者用户 ID。
        created_at:
          type: string
          format: date-time
          title: Created At
          description: 创建时间。
        updated_at:
          type: string
          format: date-time
          title: Updated At
          description: 最近更新时间。
      type: object
      required:
      - id
      - project_id
      - resource_type
      - category
      - name
      - content
      - version
      - created_by
      - updated_by
      - created_at
      - updated_at
      title: ResourceResponse
      description: 项目资源及其版本和审计字段。
    ResourceUpdateRequest:
      properties:
        expected_version:
          type: integer
          minimum: 1.0
          title: Expected Version
          description: 最近读取的资源版本；过期时返回 409。
          examples:
          - 1
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
          description: 更新后的文件名。
          examples:
          - support-agent
        content:
          type: string
          maxLength: 100000
          title: Content
          description: 更新后的正文，最多 100,000 字符。
          examples:
          - 你是一个专业的客服助手。回答时保持简洁。
      type: object
      required:
      - expected_version
      - name
      - content
      title: ResourceUpdateRequest
      description: 使用乐观锁更新资源名称与正文。
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer